+++ R/default_params_doc.R
@@ -46,6 +46,8 @@
 #' @param example_filename name of the example file
 #' @param fam_filename name of a \code{PLINK} \code{.fam} file
 #' Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.
+#' @param lmiss_filename name of a \code{PLINK} \code{.lmiss} file
+#' Use \link{read_plink_lmiss_file} to read a \code{PLINK} \code{.lmiss} file.
 #' @param log_filename name of a \code{PLINK} \code{.log} file
 #' @param maf minor allele frequency threshold.
 #' Alleles that have a frequency lower than the MAF
@@ -229,6 +231,7 @@ default_params_doc <- function(
   epistatic_phenotype_value,
   example_filename,
   fam_filename,
+  lmiss_filename,
   log_filename,
   maf,
   mafs,