+++ NEWS.md
@@ -9,6 +9,7 @@ Newest versions at top.
  * Add vignette with PLINK tutorial
  * Add `install_plink_tutorial_data` and `uninstall_plink_tutorial_data`
    to install and uninstall the PLINK tutorial data
+ * Add `read_plink_lmiss_file` to read PLINK `.lmiss` files
 
 ### MINOR IMPROVEMENTS